# packages/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Persona API",
    "summary": "Hosted service for building and running typed-memory AI personas.",
    "version": "0.8.0"
  },
  "paths": {
    "/v1/personas": {
      "post": {
        "tags": ["personas"],
        "summary": "Create Persona",
        "description": "Create a persona from YAML; populate memory stores; auto-generate an avatar.\n\nThe avatar is generated only when the builder supplied none (D-29-3); the\ngeneration is fail-soft so create never fails on an imagegen problem\n(D-29-X-fail-soft). A user-supplied ``avatar_url`` always wins (criterion 6).",
        "operationId": "create_persona_v1_personas_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePersonaRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["personas"],
        "summary": "List Personas",
        "description": "List the caller's personas (paginated; RLS-scoped).",
        "operationId": "list_personas_v1_personas_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonaSummary"
                  },
                  "title": "Response List Personas V1 Personas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/personas/{persona_id}/imagegen": {
      "post": {
        "tags": ["imagegen"],
        "summary": "Post Imagegen",
        "description": "Generate one or more images for a persona; persist to workspace; audit.\n\nReturns 201 + an ImageRef-shape payload on success. The bytes land at\n``{workspace_root}/{user_id}/{persona_id}/uploads/{blake2b}.{ext}``\n(D-13-4 layout reused per D-15-X-workspace-coordination) and are\nfetched via the existing ``GET /v1/personas/:id/uploads/:ref`` route.\n\nArgs:\n    persona_id: Persona id from the path; pre-flight RLS-checked so\n        cross-tenant ids surface as 404.\n    body: Validated :class:`ImageGenRequest` (Pydantic 422 on shape\n        violations before we get here).\n    request: FastAPI request \u2014 used for ``app.state`` access to the\n        RLS engine, the workspace root, and the composed image backend.\n    user: Authenticated principal from the bearer token.\n\nReturns:\n    ImageRef-shape JSON payload (mirrors the Spec 13 upload response\n    but as a list since ``count`` can be 2).\n\nRaises:\n    ImageGenUnavailableError: The provider is not configured at all\n        (no ``PERSONA_IMAGEGEN_API_KEY`` at startup) \u2192 503 via the\n        app exception handler.\n    HTTPException: 502 for upstream provider failures; 422 for\n        content rejection (provider moderation or hard-line filter);\n        403 for ``ToolNotAllowedError``; 404 for cross-tenant; 402 for\n        credits exhaustion; 429 for concurrency cap or rate limit.",
        "operationId": "post_imagegen_v1_personas__persona_id__imagegen_post",
        "parameters": [
          {
            "name": "persona_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Persona Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageGenRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Post Imagegen V1 Personas  Persona Id  Imagegen Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePersonaRequest": {
        "properties": {
          "yaml": {
            "type": "string",
            "title": "Yaml"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["yaml"],
        "title": "CreatePersonaRequest",
        "description": "Create a persona from a YAML document (validated against the v1.0 schema).\n\n``avatar_url`` is an optional presentation field (not part of the YAML\nschema) \u2014 the persona-list / chat-header visual identity."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImageGenRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "minLength": 1,
            "title": "Prompt"
          },
          "size": {
            "type": "string",
            "enum": ["1024x1024", "1024x1792", "1792x1024"],
            "title": "Size",
            "default": "1024x1024"
          },
          "count": {
            "type": "integer",
            "maximum": 2.0,
            "minimum": 1.0,
            "title": "Count",
            "default": 1
          },
          "quality": {
            "type": "string",
            "enum": ["standard", "high"],
            "title": "Quality",
            "default": "standard"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["prompt"],
        "title": "ImageGenRequest",
        "description": "Body of ``POST /v1/personas/:id/imagegen``.\n\nThe closed Literal surface on ``size`` and ``quality`` ensures invalid\nvalues land as 422 Pydantic validation errors before any service-\nlayer work happens; ``count`` is bounded by D-15-3 (``le=2``).\n\nAttributes:\n    prompt: The user-supplied text prompt. Required, min length 1.\n        The visual_style merge runs at the service layer\n        (:func:`persona_api.imagegen.service.generate`) so the prompt\n        here is the raw user input \u2014 NOT yet merged.\n    size: One of the three closed presets per D-15-3. Defaults to\n        ``\"1024x1024\"``. The OpenAI backend rounds non-square presets\n        per D-15-X-size-rounding; the audit captures the REQUESTED\n        value (this field's literal), not the rounded one.\n    count: Number of images to generate. ``Field(ge=1, le=2)`` enforces\n        the D-15-3 cap. Defaults to 1.\n    quality: One of the two closed presets per D-15-3. Defaults to\n        ``\"standard\"``."
      },
...
      "PersonaCapabilities": {
        "properties": {
          "vision": {
            "type": "boolean",
            "title": "Vision"
          },
          "configured_tiers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Configured Tiers"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["vision", "configured_tiers"],
        "title": "PersonaCapabilities",
        "description": "Deployment-derived capability flags surfaced with the persona detail.\n\nHydrated from the runtime :class:`persona_runtime.tier.TierRegistry` so the\nUI can answer \"does this persona support image attachments?\" BEFORE the\nuser attempts to send (Spec 13 fail-loud made visible \u2014 Spec F3 \u00a710 #7;\nD-F3-X-no-vision-surface-shape). At v0.1 the answer is deployment-wide:\nevery persona under a given deployment shares the same registry, so\n``vision`` is identical across personas \u2014 see D-F3-X-deployment-vs-persona-\ncapability-framing. The field's shape survives the v0.2 inflection where\nper-persona tier pins make the answer genuinely per-persona; only the\nhydration source changes (from registry to per-persona lookup).\n\nAttributes:\n    vision: ``True`` iff at least one configured tier resolves to a\n        backend whose ``supports_vision`` is ``True``. Read via the\n        public :meth:`TierRegistry.supports_vision_for` method\n        (D-F3-X-tier-registry-public-contract).\n    configured_tiers: Tier names registered on the active deployment\n        in insertion order (``(\"small\", \"mid\", \"frontier\")`` for the\n        typical three-tier deployment). The UI may surface these in a\n        disabled-attach tooltip to explain *which* models the deployment\n        has configured."
      },
      "PersonaDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "yaml": {
            "type": "string",
            "title": "Yaml"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "capabilities": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PersonaCapabilities"
              },
              {
                "type": "null"
              }
            ]
          },
          "consent_to_auto_dispatch": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent To Auto Dispatch"
          },
          "consent_updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Updated At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "yaml",
          "schema_version",
          "created_at",
          "updated_at"
        ],
        "title": "PersonaDetail",
        "description": "A persona's full detail (YAML + metadata).\n\nThe optional :attr:`capabilities` field (D-F3-X-capability-endpoint) is\nadditive on top of the Spec 08 / Spec 09 surface: tests + composition\nroots that do not wire a :class:`TierRegistry` (e.g. unit fixtures\nwithout the runtime) omit the field and the API returns ``None`` so the\npersona-detail surface stays usable without runtime composition."
      },
      "PersonaSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["id", "name", "role", "created_at", "updated_at"],
        "title": "PersonaSummary",
        "description": "A persona in a list view (no full YAML)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}